Create .powershell-psscriptanalyzer.psd1
Created some exceptions to the PowerShell lint program. I couldn't add them to the source itself, because it was a file to be included as a "userdata" for an EC2 instance and therefore it had to have <powershell> as the first line. This caused the PowerShell lint to ignore the in source directives.